Build phram support in all the time
i.e. Remove the derivation/Makefile target that builds an image without it.
It adds hardly anything to image size and if you're doing an upgrade using
kexec there seems to be no point in running one phram-enabled kernel for
testing then flashing a *different* image when you're happy it works.  You
should be running the tested code, not some other image that you only _hope_ is
closely related to it

This may tip the trendnet tew712 image over the magic 4MB threshold again,
but I don't care, that's a special case and we can treat it specially.