set monit.id in the image: faster boot
Switch to using /etc/monit.id (pregenerated based on host key) instead
of /run (created on the device at startup time).

Previously it took a long time for monit to start because it wants to
generate a random string for /run/monit.id and (I am guessing) there
is no entropy in the pool.  Sometimes it even starts and then writes a
zero-length idfile, which means it doesn't work properly.