check for MIPS in overlay
We've overriden deriviations for e.g. coreutils and git to make them
smaller, but these overlays apply to "build" (x86-64) packages as well
as "host" (mips) packages, so if we want to depend on x86-64 things
(like qemu) we have to do an awful lot of rebuilding

Checking for MIPS is not the right answer ( what if your build box was
mips? or your target was arm?) but it's a workaround for the time being