add config to services
Calling service.mergedConfig will return an attrset of
the config for this service merged with all the config for
all of its dependencies

This is so that services can describe e.g. busybox applets or kernel
config that they require. The data is not used directly by the
services, but will make it possible to write a module that
includes all the stuff